(split)I made all these changes on my own working copy until i had something really usable wihtout regressions for most users.
I think its place is now in CVS. Here is a ChangeLog summary
generated from my local patches.

base-0
    Imported xvidcore 0.9.0 into arch repository
patch-1
    Updated to current stable CVS_HEAD
patch-2
    Changed build system for Unix OSes
patch-3
    Fixes for the new build system in sources.
patch-4
    Fixed options and added the --disable-assembly option
patch-5
    Fixed a BSD checking in nasm output format.
patch-6
    Fixed the "ar" "s" option for some platforms.
patch-7
    Changed the way I add strings into variables.
patch-8
    Fixed portab.h for _DEBUG target and GCC/ICC compilers.
patch-9
    Added Altivec detection (Guillaume Morin)
patch-10
    Fixed MacOSX build.
patch-11
    Fixed WIN32/_MSC_VER confusion and updated MSVC libxvidcore.dsp project file.
patch-12
    DivX4 compatibility layer has been turned into an option (default:disable).

ToDo:
    - Fix IA64 assembly file selection.
    - Do I disable PPC assembly code ? (according to Guillaume, it needs to be updated)


git-svn-id: http://svn.xvid.org/trunk@824 060a26f8-e6a2-0410-8959-474133d486e0

+++ src/bitstream/bitstream.h
@@ -50,7 +50,7 @@
  *  exception also makes it possible to release a modified version which
  *  carries forward this exception.
  *
- * $Id: bitstream.h,v 1.15 2002-11-26 23:44:10 edgomez Exp $
+ * $Id: bitstream.h,v 1.16 2003-02-09 19:32:52 edgomez Exp $
  *
  ****************************************************************************/
 
@@ -301,7 +301,7 @@ BitstreamByteAlign(Bitstream * const bs)
 static uint32_t __inline
 BitstreamPos(const Bitstream * const bs)
 {
-	return 8 * ((ptr_t)bs->tail - (ptr_t)bs->start) + bs->pos;
+	return((uint32_t)(8*((ptr_t)bs->tail - (ptr_t)bs->start) + bs->pos));
 }
 
 
@@ -312,7 +312,7 @@ BitstreamPos(const Bitstream * const bs)
 static uint32_t __inline
 BitstreamLength(Bitstream * const bs)
 {
-	uint32_t len = (ptr_t) bs->tail - (ptr_t) bs->start;
+	uint32_t len = (uint32_t)((ptr_t)bs->tail - (ptr_t)bs->start);
 
 	if (bs->pos) {
 		uint32_t b = bs->buf;
